Seperated the STJ logic from the JsonPackageSpec reader
